First version of egg breeding analysis for gens 2 to 5
The logic is given a species pokemon and egg moves learned by the pokemon find a valid egg chain that allow this combination to be legal.
For that if the species could have a father that can learn all the moves without using egg moves, if not check every potential father that can learn the moves including egg moves and recursively find for that species a father that can learn all the moves.
Eventualy if the combination is legal the chain will end with an species that do not needs egg moves, like for example species A with 3 egg moves, found a father B who learns 2 moves with level up and 1 by egg and found a species C father of B that learn the last move by TM.

To make it speed the list of potential fathers and fathers by egg group is precomputed at the start of the program. The list of eggmoves learned by the fathers is precomputed too

There is many things not done yet: special pokemon like events and static encounters that can inherit their special moves, cross generation breeding, check max possible depth of legal combinations, gen 1 moves in gen 2 breeding,..